Revert "refactor: keep metadata query in file slice"
This reverts commit 5c99cc45512560499247b168aea0a2c4b39e2030.
