+++ apps/web/src/components/breadcrumbs/pdf-breadcrumb.tsx
@@ -4,7 +4,7 @@ import { Link, useMatch } from "@tanstack/react-router";
 import { BidiText } from "@stll/ui/components/bidi-text";
 import { BreadcrumbItem } from "@stll/ui/components/breadcrumb";
 
-import { fileMetadataOptions } from "@/routes/_protected.workspaces/$workspaceId/-components/files/queries";
+import { fileMetadataOptions } from "@/lib/files/file-metadata-query";
 
 export const PdfBreadcrumb = () => {
   const pdfMatch = useMatch({
